+++ components/pages/Home.tsx
@@ -1,10 +1,25 @@
 import React from "react";
 import { StyleSheet, TextInput } from "react-native";
 import { createStackNavigator } from "@react-navigation/stack";
+import { createMaterialTopTabNavigator } from "@react-navigation/material-top-tabs";
 import Icon from "react-native-vector-icons/FontAwesome";
-import { Products } from "../atoms/Home";
+import { Category, Products } from "../atoms/Home";
 
+const Tab = createMaterialTopTabNavigator();
 const Stack = createStackNavigator();
+
+function SearchTab({ value }) {
+  return (
+    <Tab.Navigator tabBarOptions={{ scrollEnabled: true }}>
+      <Tab.Screen name="おすすめ">
+        {(props) => <Products {...props} value={value} />}
+      </Tab.Screen>
+      <Tab.Screen name="新着" component={Products} />
+      <Tab.Screen name="カテゴリー" component={Category} />
+      <Tab.Screen name="保存した検索条件" component={Products} />
+    </Tab.Navigator>
+  );
+}
 
 export default function Home() {
   const [value, setValue] = React.useState("");
@@ -13,7 +28,6 @@ export default function Home() {
     <Stack.Navigator>
       <Stack.Screen
         name="Home"
-        component={Products}
         options={{
           headerTitle: () => (
             <TextInput
@@ -22,6 +36,7 @@ export default function Home() {
                 setValue(newValue);
               }}
               placeholder="検索"
+              autoCapitalize="none"
             />
           ),
           headerLeft: () => <Icon name="qrcode" size={24} />,
@@ -29,7 +44,9 @@ export default function Home() {
           headerLeftContainerStyle: styles.headerLeft,
           headerRightContainerStyle: styles.headerRight,
         }}
-      />
+      >
+        {(props) => <SearchTab value={value} />}
+      </Stack.Screen>
     </Stack.Navigator>
   );
 }
